feat: adjust sidebar item widths, alignment
Made all of the menu items to be flex items so that the image and text
could be aligned vertically.

Adjusted nav menu grid col width so that it would only be as long as the
longest item. This makes the link regions more reasonable.

Made branding title text to be bold font weight.